organize test files in a directory
